refactor: deconstruct monolithic classroom socket.js into modular sub-handlers
